Refactoring the UserAccountController into two controllers and added respective Unit & Integration tests
