--- UserApi/UserApi.IntegrationTests/Controllers/UserAccountControllerTests.cs
@@ -1,232 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Net;
-using System.Net.Http;
-using System.Net.Http.Headers;
-using System.Text;
-using NUnit.Framework;
-using System.Threading.Tasks;
-using UserApi.Contracts.Requests;
-using UserApi.Contracts.Responses;
-using FluentAssertions;
-using NUnit.Framework.Constraints;
-using Testing.Common.Helpers;
-using UserApi.Services.Models;
-
-namespace UserApi.IntegrationTests.Controllers
-{
-    public class UserAccountControllerTests  : ControllerTestsBase
-    {
-        private readonly UserAccountEndpoints _userAccountEndpoints = new ApiUriFactory().UserAccountEndpoints;
-        private string _newUserId;
-        
-        [Test]
-        public async Task should_get_group_by_name_not_found_with_bogus_group_name()
-        {
-            var groupName = "foo";
-            var getResponse = await SendGetRequestAsync(_userAccountEndpoints.GetGroupByName(groupName));
-            getResponse.StatusCode.Should().Be(HttpStatusCode.NotFound);
-        }
-
-        [Test]
-        public async Task should_get_group_by_name()
-        {
-            var groupName = "SSPR Enabled";
-            var getResponse = await SendGetRequestAsync(_userAccountEndpoints.GetGroupByName(groupName));
-            getResponse.StatusCode.Should().Be(HttpStatusCode.OK);
-            var groupResponseModel =
-                ApiRequestHelper.DeserialiseSnakeCaseJsonToResponse<GroupsResponse>(getResponse.Content
-                    .ReadAsStringAsync().Result);
-            groupResponseModel.DisplayName.Should().Be(groupName);
-            groupResponseModel.GroupId.Should().NotBeNullOrWhiteSpace();
-        }
-        
-        [Test]
-        public async Task should_get_group_by_id()
-        {
-            var groupId = "8881ea85-e0c0-4a0b-aa9c-979b9f0c05cd";
-            var getResponse = await SendGetRequestAsync(_userAccountEndpoints.GetGroupById(groupId));
-            getResponse.StatusCode.Should().Be(HttpStatusCode.OK);
-            var groupResponseModel =
-                ApiRequestHelper.DeserialiseSnakeCaseJsonToResponse<GroupsResponse>(getResponse.Content
-                    .ReadAsStringAsync().Result);
-            groupResponseModel.GroupId.Should().Be(groupId);
-        }
-        
-        [Test]
-        public async Task should_get_group_by_id_not_found_with_bogus_id()
-        {
-            var groupId = Guid.Empty.ToString();
-            var getResponse = await SendGetRequestAsync(_userAccountEndpoints.GetGroupById(groupId));
-            getResponse.StatusCode.Should().Be(HttpStatusCode.NotFound);
-        }
-
-        [Test]
-        public async Task should_get_user_by_id_not_found_with_bogus_user_id()
-        {
-            var userId = "foo";
-            var getResponse = await SendGetRequestAsync(_userAccountEndpoints.GetUserByAdUserId(userId));
-            getResponse.StatusCode.Should().Be(HttpStatusCode.NotFound);
-        }
-
-        [Test]
-        public async Task should_get_user_by_id()
-        {
-            const string userId = "84fa0832-cd70-4788-8f48-e869571e0c56";
-            var getResponse = await SendGetRequestAsync(_userAccountEndpoints.GetUserByAdUserId(userId));
-            getResponse.StatusCode.Should().Be(HttpStatusCode.OK);
-            var userResponseModel =
-                ApiRequestHelper.DeserialiseSnakeCaseJsonToResponse<UserDetailsResponse>(getResponse.Content
-                    .ReadAsStringAsync().Result);
-            userResponseModel.UserId.Should().Be(userId);
-            userResponseModel.Username.Should().NotBeNullOrWhiteSpace();
-            userResponseModel.DisplayName.Should().NotBeNullOrWhiteSpace();
-        }
-        
-        [Test]
-        public async Task should_get_user_by_recovery_email()
-        {
-            const string alternativeEmail = "judge@kinly.com";
-            var getResponse = await SendGetRequestAsync(_userAccountEndpoints.GetUserByRecoveryEmail(alternativeEmail));
-            getResponse.StatusCode.Should().Be(HttpStatusCode.OK);
-            var userResponseModel =
-                ApiRequestHelper.DeserialiseSnakeCaseJsonToResponse<UserDetailsResponse>(getResponse.Content
-                    .ReadAsStringAsync().Result);
-            userResponseModel.UserId.Should().NotBeNullOrWhiteSpace();
-            userResponseModel.Username.Should().NotBeNullOrWhiteSpace();
-            userResponseModel.DisplayName.Should().NotBeNullOrWhiteSpace();
-        }
-        
-        [Test]
-        public async Task should_get_user_by_recovery_email_not_found_with_bogus_mail()
-        {
-            const string alternativeEmail = "i.do.not.exist@nowhere.ever.com";
-            var getResponse = await SendGetRequestAsync(_userAccountEndpoints.GetUserByRecoveryEmail(alternativeEmail));
-            getResponse.StatusCode.Should().Be(HttpStatusCode.NotFound);
-        }
-
-        [Test]
-        public async Task should_create_citizen_user_on_ad()
-        {
-            var createUserRequest = new CreateUserRequest
-            {
-                RecoveryEmail = Faker.Internet.Email(),
-                FirstName = Faker.Name.First(),
-                LastName = Faker.Name.Last()
-            };
-            var createUserHttpRequest = new StringContent(
-                ApiRequestHelper.SerialiseRequestToSnakeCaseJson(createUserRequest),
-                Encoding.UTF8, "application/json");
-
-            var createUserResponse =
-                await SendPostRequestAsync(_userAccountEndpoints.CreateUser, createUserHttpRequest);
-
-            createUserResponse.StatusCode.Should().Be(HttpStatusCode.Created);
-            var createUserModel =
-                ApiRequestHelper.DeserialiseSnakeCaseJsonToResponse<NewUserResponse>(createUserResponse.Content
-                    .ReadAsStringAsync().Result);
-            TestContext.WriteLine($"Response:{ApiRequestHelper.SerialiseRequestToSnakeCaseJson(createUserModel)}");
-            createUserModel.Should().NotBeNull();
-            createUserModel.UserId.Should().NotBeNullOrEmpty();
-            createUserModel.Username.ToLower().Should()
-                .Be($@"{createUserRequest.FirstName}.{createUserRequest.LastName}@hearings.reform.hmcts.net".ToLower());
-            createUserModel.OneTimePassword.Should().NotBeNullOrEmpty();
-
-            _newUserId = createUserModel.UserId;
-
-            var addSsprGroupRequest = new AddUserToGroupRequest
-                {UserId = createUserModel.UserId, GroupName = "SSPR Enabled"};
-            var addSsprGroupHttpRequest = new StringContent(
-                ApiRequestHelper.SerialiseRequestToSnakeCaseJson(addSsprGroupRequest),
-                Encoding.UTF8, "application/json");
-            var addSsprGroupHttpResponse =
-                await SendPatchRequestAsync(_userAccountEndpoints.AddUserToGroup, addSsprGroupHttpRequest);
-            addSsprGroupHttpResponse.IsSuccessStatusCode.Should().BeTrue();
-
-            var addExternalGroupRequest = new AddUserToGroupRequest
-                {UserId = createUserModel.UserId, GroupName = "External"};
-            var addExternalGroupHttpRequest = new StringContent(
-                ApiRequestHelper.SerialiseRequestToSnakeCaseJson(addExternalGroupRequest),
-                Encoding.UTF8, "application/json");
-            var addExternalGroupHttpResponse =
-                await SendPatchRequestAsync(_userAccountEndpoints.AddUserToGroup, addExternalGroupHttpRequest);
-            addExternalGroupHttpResponse.IsSuccessStatusCode.Should().BeTrue();
-        }
-        
-        [Test]
-        public async Task should_get_groups_for_user()
-        {
-            const string userId = "84fa0832-cd70-4788-8f48-e869571e0c56";
-            var getResponse = await SendGetRequestAsync(_userAccountEndpoints.GetGroupsForUser(userId));
-            getResponse.StatusCode.Should().Be(HttpStatusCode.OK);
-            var groupsForUserModel =
-                ApiRequestHelper.DeserialiseSnakeCaseJsonToResponse<List<GroupsResponse>>(getResponse.Content
-                    .ReadAsStringAsync().Result);
-            groupsForUserModel.Should().NotBeEmpty();
-        }
-
-        [Test]
-        public async Task should_get_user_profile_by_email()
-        {
-            const string email = "VirtualRoomAdministrator@kinley.com";
-            var getResponse = await SendGetRequestAsync(_userAccountEndpoints.GetUserProfileByEmail(email));
-            getResponse.StatusCode.Should().Be(HttpStatusCode.OK);
-            var userResponseModel =
-                ApiRequestHelper.DeserialiseSnakeCaseJsonToResponse<UserProfile>(getResponse.Content
-                    .ReadAsStringAsync().Result);
-            userResponseModel.UserName.Should().NotBeNullOrWhiteSpace();
-            userResponseModel.Email.Should().NotBeNullOrWhiteSpace();
-            userResponseModel.FirstName.Should().NotBeNullOrWhiteSpace();
-            userResponseModel.LastName.Should().NotBeNullOrWhiteSpace();
-            userResponseModel.UserRole.Should().NotBeNullOrWhiteSpace();
-        }
-
-        [Test]
-        public async Task should_get_user_profile_by_user_principal_name()
-        {
-            const string username = "VirtualRoomAdministrator@hearings.reform.hmcts.net";
-            var getResponse = await SendGetRequestAsync(_userAccountEndpoints.GetUserProfileByUserName(username));
-            getResponse.StatusCode.Should().Be(HttpStatusCode.OK);
-            var userResponseModel =
-                ApiRequestHelper.DeserialiseSnakeCaseJsonToResponse<UserProfile>(getResponse.Content
-                    .ReadAsStringAsync().Result);
-            userResponseModel.UserName.Should().NotBeNullOrWhiteSpace();
-            userResponseModel.Email.Should().NotBeNullOrWhiteSpace();
-            userResponseModel.FirstName.Should().NotBeNullOrWhiteSpace();
-            userResponseModel.LastName.Should().NotBeNullOrWhiteSpace();
-            userResponseModel.UserRole.Should().NotBeNullOrWhiteSpace();
-        }
-
-        [Test]
-        public async Task should_get_profile_by_email_not_found_with_bogus_mail()
-        {
-            const string email = "i.do.not.exist@nowhere.ever.com";
-            var getResponse = await SendGetRequestAsync(_userAccountEndpoints.GetUserProfileByEmail(email));
-            getResponse.StatusCode.Should().Be(HttpStatusCode.NotFound);
-        }
-
-        [Test]
-        public async Task should_get_user_profile_by_user_principal_name_not_found_with_bogus_mail()
-        {
-            const string username = "i.do.not.exist@hearings.reform.hmcts.net";
-            var getResponse = await SendGetRequestAsync(_userAccountEndpoints.GetUserProfileByUserName(username));
-            getResponse.StatusCode.Should().Be(HttpStatusCode.NotFound);
-        }
-
-        [TearDown]
-        public void ClearUp()
-        {
-            if (string.IsNullOrWhiteSpace(_newUserId)) return;
-            using (var client = new HttpClient())
-            {
-                client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", GraphApiToken);
-                var httpRequestMessage = new HttpRequestMessage(HttpMethod.Get,
-                    $@"https://graph.microsoft.com/v1.0/users/{_newUserId}");
-                var result = client.SendAsync(httpRequestMessage).Result;
-                result.IsSuccessStatusCode.Should().BeTrue($"{_newUserId} should be deleted");
-                _newUserId = null;
-            }
-        }
-
-    }
-}
